refactor(list-categories-usecase): move to search package
